Variable array API
This commit adds a new representation to Dr.Jit that groups a number of
regular variables into a "variable array". The API exposes convenient
functionality to read and write elements of this variable array.
Pre-initialized arrays storing a literal constant can also be created.

The main difference between variable arrays and a collection of regular
variables is that this representation supports reads/writes using
computed indices. This enables previously impossible use cases of Dr.Jit.

At an implementation level, array variables are stored on the
stack---for example, the LLVM version uses ``alloca()`` to reserve
memory within the function scope.  That said, it is possible that
optimization passes in the backend compiler might turn this
stack-allocated memory back to CPU/GPU registers.

+++ src/cuda_tex.cpp
@@ -540,7 +540,7 @@ void jitc_cuda_tex_lookup(size_t ndim, const void *texture_handle,
 
     for (size_t ti = 0; ti < tex.n_textures; ++ti) {
         // Perform a fetch per texture ..
-        v.kind = VarKind::TexLookup;
+        v.kind = (uint32_t) VarKind::TexLookup;
         memset(v.dep, 0, sizeof(v.dep));
         const Variable *active_v = jitc_var(active);
         if (active_v->is_literal() && active_v->literal == 1) {
@@ -560,7 +560,7 @@ void jitc_cuda_tex_lookup(size_t ndim, const void *texture_handle,
         Ref tex_load = steal(jitc_var_new(v));
 
         // .. and then extract components
-        v.kind = VarKind::Extract;
+        v.kind = (uint32_t) VarKind::Extract;
         memset(v.dep, 0, sizeof(v.dep));
         for (size_t ch = 0; ch < tex.channels(ti); ++ch) {
             v.literal = (uint64_t) ch;
@@ -584,7 +584,7 @@ void jitc_cuda_tex_bilerp_fetch(size_t ndim, const void *texture_handle,
     for (size_t ti = 0; ti < tex.n_textures; ++ti) {
         for (size_t ch = 0; ch < tex.channels(ti); ++ch) {
             // Perform a fetch per texture and channel..
-            v.kind = VarKind::TexFetchBilerp;
+            v.kind = (uint32_t) VarKind::TexFetchBilerp;
             v.literal = ch;
             memset(v.dep, 0, sizeof(v.dep));
             v.dep[0] = tex.indices[ti];
@@ -598,7 +598,7 @@ void jitc_cuda_tex_bilerp_fetch(size_t ndim, const void *texture_handle,
             Ref tex_load = steal(jitc_var_new(v));
 
             memset(v.dep, 0, sizeof(v.dep));
-            v.kind = VarKind::Extract;
+            v.kind = (uint32_t) VarKind::Extract;
             for (uint32_t j = 0; j < 4; ++j) {
                 // .. and then extract components
                 v.literal = (uint64_t) j;
